requirements.txt
-e .
pyyaml==5.3.1
requests==2.25.1